<!-- Thank you for contributing to openGemini! -->

<!-- Mark the checkbox [X] or [x] if you agree with the item. -->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close","fix", "resolve" or "ref".
-->

Issue Number: close/fix/resolve/ref #383 

### What is changed and how it works?
现在实现了在运行时统计不同tag filter影响的时间线数量，在查询时先将语句所有的tag filter提取出来，然后根据时间线数量排序，时间线数量少的tag filter排在前面。首先将排在最前面的tag的所有时间线的集合作为结果集合的基础，遍历之后的tag，如果下一个tag的时间线数量/当前集合大小<=10,那么就取出这个tag的所有时间线，与当前集合做交集；否则的话，认为遍历+求交集的代价大于剪枝的代价，进行剪枝：遍历当前结果集合，判断某个元素是否符合剩下所有的tag filter的要求，如果符合，则保留，否则将删除。

### How Has This Been Tested?
提供了一个脚本，向testdb库test_01表中注入100万条数据，数据有4个tag列和1个field列，4个tag列的基数分别是：1，100，10000，1000000。我运行了这个脚本5次，表中一共是500万条数据。
测试的查询语句为：
```sql
select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
```
优化之前的测试结果：
```
mxk@XINKAIMA-NB1:~/openGemini-copy$ ./build/ts-cli
2023/10/22 17:20:48 proto: duplicate proto type registered: influxql.Measurements
2023/10/22 17:20:48 proto: duplicate proto type registered: influxql.Measurement
openGemini CLI 0.1.0 (rev-revision)
Please use `quit`, `exit` or `Ctrl-D` to exit this program.
> use testdb
> timer
Timer is enabled
Elapsed: 56.6µs
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 82.0419ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 75.6909ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 38.4417ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 86.3376ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 41.166ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 35.8974ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 34.5085ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 83.9511ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 32.5448ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 30.2746ms
> 
```
稳定之后大概是30ms左右，对优化后的版本进行测试：
```
mxk@XINKAIMA-NB1:~/openGemini-copy$ ./build/ts-cli
2023/10/22 17:18:12 proto: duplicate proto type registered: influxql.Measurements
2023/10/22 17:18:12 proto: duplicate proto type registered: influxql.Measurement
openGemini CLI 0.1.0 (rev-revision)
Please use `quit`, `exit` or `Ctrl-D` to exit this program.
> use testdb
> timer
Timer is enabled
Elapsed: 55.9µs
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 77.0088ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 9.0707ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 6.4436ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 14.4037ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 3.9671ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 4.8968ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 4.268ms
> select count(*) from test_01 where tag0='t0' and tag1='t1_88' and tag2='t2_8848' and tag3='t3_884848'
name: test_01
+------+-------------+
| time | count_value |
+------+-------------+
| 0    | 5           |
+------+-------------+
2 columns, 1 rows in set

Elapsed: 4.1166ms
> 
```
稳定下来后，耗时大概在4ms之间，实现了7倍以上的性能提升！  
根据分析，查询变快的关键是进行剪枝。比如说，现在的剪枝阈值是loopsCount/currentSetSize >= 10，在当前的测试语句中，会按照tag3->tag2->tag1->tag0的顺序进行遍历，首先选取tag3的时间线集合为基础，集合中有1个元素，然后遍历tag2，tag2中有100条符合要求的时间线，此时进行剪枝，只需要对当前集合中的那一条时间线进行判断即可，不需要遍历tag2索引到的100条时间线；遍历到tag1时，tag1有10000条符合要求的时间线，此时继续剪枝，并且剪枝的收益是巨大的。
总结一下，当前优化方案的工作流是：
1. 某条sql语句第一次查询时，此时不知道它的tag filter的信息，每一个tagFilter的代价都被估计成0，在查询的过程中维护map，map的key是tagFilter，value是查询到的时间线数量
2. 再次查询时，就可以根据tagFilter索引到的时间线数量，对tagFilter进行排序，时间线数量最少的tagFilter放在前面。
3. 查询第一个tagFilter所有的时间线，作为结果集合的初始值
4. 遍历剩余的tagFilter，如果它的时间线数量小于 当前的结果集合大小乘以某个阈值，那么就查询这个tagFilter的所有时间线，与结果集合取交集，并维护map。否则的话，就进行剪枝，逐个判断结果集合中的元素是否符合剩余tagFilter的要求，如果不符合，则删去。

使用explain analyze语句进行分析，可以看出，开启剪枝前，索引查询时间为：
```
tsid_search:pp=65.770601ms
```
开启剪枝后，时间缩短为：
```
tsid_search:pp=1.267801ms
```
这一部分的提升是非常显著的。

数据注入的脚本：
```python
# inject 100w points
import requests
import random
points = []
mst = 'test_01'
for i in range(100):
    for j in range(100):
        for k in range(100):
            point = f'{mst},tag0=t0,tag1=t1_{i},tag2=t2_{i*100+j},tag3=t3_{i*10000+j*100+k} value={random.random()}'
            points.append(point)

headers = {
    "Content-Type": "text/plain"
}
url = 'http://127.0.0.1:8086/write?db=testdb'

for i in range(1000):
    data = '\n'.join(points[i*1000:(i+1)*1000])
    for j in range(5):
        response = requests.post(url, data=data, headers=headers)
```
# Checklist:

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules